affichage /login

+++ src/app/src/app/UserBack/user-list/user-list.component.ts
@@ -5,6 +5,7 @@ import { ngxCsv } from 'ngx-csv';
 import { AccorService } from 'src/app/accor.service';
 import { User } from '../user';
 import { Location } from '@angular/common'
+import { checkServerIdentity } from 'tls';
 
 
 @Component({
@@ -15,37 +16,45 @@ import { Location } from '@angular/common'
 export class UserListComponent implements OnInit {
 
   @ViewChild("dispatcher") myNameElem!: ElementRef;
-  val = localStorage.getItem("dispatcher");
+  val = localStorage.getItem('dispatcher');
 
+  
   @Input() pdata:any;
-
+  
   private _fb: any;
   form: any;
   type = 'Head of Department';
   spend_limit = 0;
   params: any;
   parameters: any;
-
+  
   branch = [
     { name: "primaryBranch" }
   ];
-
+  
   selectedCompanies = [];
-
+  
   users: any;
   tabUser: any = [];
   tabcc: any = [];
   searchKey: string = "";
   searchTerm: string = "";
-
+  
   displayStyle = "none";
   displayStyle2 = "none";
   displayHome = 'none';
-  tableStyle = "block";
-
-  dispatcher = true;
+  tableStyle = "table";
+  
+  dispatcher :any;
   t:any;
   tabi : any = [];
+
+  getDataHmc = localStorage.getItem('getDataHmc');
+  getDataHn = localStorage.getItem('getDataHn');
+  getDataGm = localStorage.getItem('getDataGm');
+  getDataBranch = localStorage.getItem('getDataBranch');
+
+
 
   userForm = new FormGroup({
     selectCompany: new FormControl,
@@ -58,39 +67,51 @@ export class UserListComponent implements OnInit {
 
   idPop:any;
 
+  role:any =[];
+  roleName:any;
+
   constructor(
     private service: AccorService,
     private router: Router,
-    private location:Location,
     public theDispatcher: AccorService
   ) { }
 
   ngOnInit(): void {
 
-    console.log(this.pdata)
+   this.getDataHmc = this.getDataHmc!.replace(/[""]/gi, '')
+   console.log(this.getDataHmc)
+   this.getDataGm = this.getDataGm!.replace(/[""]/gi, '')
+   console.log(this.getDataGm)
+   this.getDataBranch = this.getDataBranch!.replace(/[""]/gi, '')
+   this.getDataHn = this.getDataHn!.replace(/[""]/gi, '')
 
-    console.log('val', this.val)
-    this.service.users()
+
+
+
+   this.service.users()
       .subscribe(data => {
         this.tabUser = data;
         console.log(this.tabUser)
         for (let i = 0; i < this.tabUser.length; i++) {
           this.tabi = this.tabUser[i];
 
-          // const arraycc = this.tabUser[i].costCenters;
-          // for (let ic = 0; ic < arraycc.length; ic++) {
-          //   this.tabcc = arraycc[ic];
-
-          //   console.log(arraycc[ic])
-          // }
-          
-          // //console.log(this.tabcc)
-          // this.tabcc
           this.t = this.tabUser[i].selectCompany
-          //console.log(this.t)
-        }
-      })
+          this.role = this.tabi.roles
 
+          for (let irole = 0; irole < this.role.length; irole++) {
+            this.role = this.role[irole];
+           // console.log(this.role)
+         
+          }
+          this.role.name
+          // console.log(this.role.name)
+          this.roleName = this.role.name
+          console.log(this.roleName)
+          console.log(this.roleName == 'ROLE_GM')
+        }
+        
+      })
+      
     this.service.search.subscribe((val: any) => {
       this.searchKey = val;
     })
@@ -100,44 +121,71 @@ export class UserListComponent implements OnInit {
         this.parameters = data;
         console.log('param', this.parameters)
       })
+
+      this.trueOrFalseHome
+    
   }
+
+
+  trueOrFalseDispatcher() {
+    if (document.getElementById('dispatcher') ) {
+      localStorage.setItem('dispatcher', JSON.stringify(this.dispatcher))
+      console.log(this.dispatcher)
+      return true
+    } else {
+      return false
+    }
+  }
+
+
 
   testDisp(dispId: any) {
-    console.log(dispId)
-          if (document.getElementById('radioCheck')!.ariaChecked) {
+          if(this.trueOrFalseDispatcher() == true){ 
             console.log(document.getElementById('radioCheck')!.ariaChecked )
-            return;
+            //document.getElementById('radioCheck')!.ariaChecked = localStorage.getItem("radio"); 
+            return dispId.selectCompany;
           } else {
             return 'null'
-          }    
+          }         
   }
 
+  save(data:any){
+    if(data != undefined){
+      localStorage.setItem('getDataDisp', JSON.stringify(data))
+    }
+    console.log(data)
 
+  }
+
+  disable(data:any){
+    var desired = this.getDataHmc!.replace(/[""]/gi, '')
+    console.log(desired)
+    
+    console.log(data.selectCompany)
+    if(desired == data.selectCompany){
+      console.log('same')
+    }
+  }
 
    openPopup(idPop:any) {
 
     console.log('_______',idPop)
    this.displayStyle = "block"
    this.tableStyle = "none"
-  
-   }
 
- 
+   }
 
   openPopup2() {
 
-    if (this.displayStyle2 = "block") {
+   this.displayStyle2 = "block"
       this.tableStyle = "none";
-      this.displayStyle = "none"
-    } else if (this.displayStyle2 = "block") {
-      this.tableStyle = "block";
-    }
+      this.displayStyle = "none"  
   }
   
   back() {
-      this.tableStyle = "block";
+      this.tableStyle = "table";;
       this.displayStyle = "none"
-
+      this.displayStyle2 = "none"
   }
 
 
@@ -159,34 +207,10 @@ export class UserListComponent implements OnInit {
     }
   }
 
-  testGM() {
-    if (this.selectedCompanies.values === this.parameters.hotel_MegaCode) {
-      console.log('test gm', this.selectedCompanies)
-      console.log('testX', this.parameters.hotel_MegaCode)
-      console.log('test gm2', this.selectedCompanies.values === this.parameters.hotel_MegaCode)
-
-      //this.parameters?.filter((p: {hotel_MegaCode: any; p: any;}) => p.hotel_MegaCode === this.selectedCompanies.values)
-
-      return this.parameters.hotel_MegaCode
-    }
-  }
-
-  recupBranch() {
-    console.log('branche', this.branch)
-    return this.branch.filter(primary => true)
-  }
-
-  trueOrFalseDispatcher() {
-    if (this.dispatcher == true) {
-      console.log(this.dispatcher)
-      return 'dispatcher'
-    } else {
-      return false
-    }
-  }
 
   trueOrFalseHome() {
-    if (this.userForm.value.primaryBranch == 'true') {
+    if (this.userForm.value.primaryBranch == true) {
+      console.log(this.userForm.value.primaryBranch)
       return 'TRUE'
     } else {
       return 'FALSE'
@@ -203,13 +227,13 @@ export class UserListComponent implements OnInit {
 
   csv(dispId: any) {
       const limit = this.approvalLimit();
-      const branche = this.recupBranch();
+      const branche = this.getDataBranch;
       const home = this.trueOrFalseHome();
-      const gm = this.testGM();
-      const disp = this.testDisp(dispId);
+      const gm= this.getDataGm;
+      //const disp = this.testDisp(dispId);
 
       const data = [
-        [branche, home, dispId.email, dispId.firstName, dispId.lastName, 'ACTIVE', gm, limit, this.spend_limit, dispId.selectCompany, 'Head of Department']
+        [branche, home, dispId.username, dispId.firstName, dispId.lastName, 'ACTIVE', gm, limit, this.spend_limit, dispId.selectCompany, 'Head of Department']
       ];
       console.log('test form', this._fb)
 
@@ -229,16 +253,18 @@ export class UserListComponent implements OnInit {
   }
 
   csvdelete(dispId: any) {
+  
     const limit = this.approvalLimit();
-    const branche = this.recupBranch();
+    const branche = this.getDataBranch;
     const home = this.trueOrFalseHome();
-    const gm = this.testGM();
+    const gm = this.getDataGm;
     const state = this.trueOrFalseState();
-    const disp = this.testDisp(dispId);
+   // const disp = this.testDisp(dispId);
 
     const data = [
-      [branche, home, dispId.email, dispId.firstName, dispId.lastName, state, gm, limit, this.spend_limit, dispId.selectCompany, 'Head of Department']
+      [branche,home , dispId.username, dispId.firstName, dispId.lastName, state, gm, limit, this.spend_limit, '', 'Head of Department']
     ];
+    
     console.log('test form', this._fb)
 
     let options = {
@@ -258,6 +284,8 @@ export class UserListComponent implements OnInit {
 
 
   remove(userId: any) {
+    // if(confirm('Continue with deletion'))
+     if(confirm('Yes I confirm, and I want to delete the user'))
     this.service.deleteUser(userId)
       .subscribe((data: any) => {
         this.users = this.users?.filter((user: { id: any; }) => userId !== user.id);
@@ -265,34 +293,6 @@ export class UserListComponent implements OnInit {
         window.location.reload()
         console.log('this.users', this.users)
       })
-
-
-
-  //   const limit = this.approvalLimit();
-  //   const branche = this.recupBranch();
-  //   const home = this.trueOrFalseHome();
-  //   const state = this.trueOrFalseState();
-  //   const gm = this.testGM();
-  //   const disp = this.testDisp(dispId);
-
-  //   const data = [
-  //     [branche, home, this.users, this.userForm.value.firstName, this.userForm.value.lastName, state, gm, limit, this.spend_limit, disp , 'Head of Department']
-  //   ];
-  //   console.log('test form', this._fb)
-
-  //   let options = {
-  //     fieldSeparator: ';',
-  //     quoteStrings: '"',
-  //     decimalseparator: '.',
-  //     showLabels: true,
-  //     showTitle: false,
-  //     useBom: true,
-  //     headers: ['BranchId', 'HOME', 'Email', 'First Name', 'Last Name', 'State', 'Manager', 'Approval limit', 'Spend_limit', 'Owned Cost Center', 'User type']
-  //   };
-  //   console.log('dataFormtoCSV', data)
-
-  //   new ngxCsv(data, "Accortemplateuserssheet", options)
-
    }
 }
